sentiment_analysis/src/serve/requirements.txt
flask
google-cloud-storage
werkzeug
python-dotenv
pandas
joblib
tensorflow
mlflow
